feat: Adjustments after review
