--- bedrock-runtime/deployment/src/test/resources/full-config.properties
@@ -1,10 +1,0 @@
-quarkus.bedrock-runtime.endpoint-override=http://localhost:9090
-
-quarkus.bedrock-runtime.aws.region=us-east-2
-quarkus.bedrock-runtime.aws.credentials.type=static
-quarkus.bedrock-runtime.aws.credentials.static-provider.access-key-id=test-key
-quarkus.bedrock-runtime.aws.credentials.static-provider.secret-access-key=test-secret
-
-quarkus.bedrock-runtime.sync-client.type = url
-quarkus.bedrock-runtime.sync-client.connection-timeout = 0.100S
-quarkus.bedrock-runtime.sync-client.socket-timeout = 0.100S
