--- bedrock-runtime/runtime/src/main/java/io/quarkiverse/amazon/bedrockruntime/runtime/BedrockRuntimeBuildTimeConfig.java
@@ -1,21 +1,0 @@
-package io.quarkiverse.amazon.bedrockruntime.runtime;
-
-import io.quarkiverse.amazon.common.runtime.DevServicesBuildTimeConfig;
-import io.quarkiverse.amazon.common.runtime.HasSdkBuildTimeConfig;
-import io.quarkiverse.amazon.common.runtime.HasTransportBuildTimeConfig;
-import io.quarkus.runtime.annotations.ConfigPhase;
-import io.quarkus.runtime.annotations.ConfigRoot;
-import io.smallrye.config.ConfigMapping;
-
-/**
- * Amazon Bedrock Runtime build time configuration
- */
-@ConfigMapping(prefix = "quarkus.bedrock-runtime")
-@ConfigRoot(phase = ConfigPhase.BUILD_AND_RUN_TIME_FIXED)
-public interface BedrockRuntimeBuildTimeConfig extends HasSdkBuildTimeConfig, HasTransportBuildTimeConfig {
-
-    /**
-     * Config for dev services
-     */
-    DevServicesBuildTimeConfig devservices();
-}
