--- bedrock-runtime/runtime/src/main/java/io/quarkiverse/amazon/bedrockruntime/runtime/BedrockRuntimeConfig.java
@@ -1,25 +1,0 @@
-package io.quarkiverse.amazon.bedrockruntime.runtime;
-
-import io.quarkiverse.amazon.common.runtime.AsyncHttpClientConfig;
-import io.quarkiverse.amazon.common.runtime.HasAmazonClientRuntimeConfig;
-import io.quarkiverse.amazon.common.runtime.SyncHttpClientConfig;
-import io.quarkus.runtime.annotations.ConfigDocSection;
-import io.quarkus.runtime.annotations.ConfigPhase;
-import io.quarkus.runtime.annotations.ConfigRoot;
-import io.smallrye.config.ConfigMapping;
-
-@ConfigRoot(phase = ConfigPhase.RUN_TIME)
-@ConfigMapping(prefix = "quarkus.bedrock-runtime")
-public interface BedrockRuntimeConfig extends HasAmazonClientRuntimeConfig {
-    /**
-     * Sync HTTP transport configurations
-     */
-    @ConfigDocSection
-    SyncHttpClientConfig syncClient();
-
-    /**
-     * Async HTTP transport configurations
-     */
-    @ConfigDocSection
-    AsyncHttpClientConfig asyncClient();
-}
