+++ bedrockruntime/deployment/src/main/java/io/quarkiverse/amazon/bedrockruntime/deployment/BedrockRuntimeProcessor.java
@@ -1,0 +1,91 @@
+package io.quarkiverse.amazon.bedrockruntime.deployment;
+
+import java.util.List;
+
+import org.jboss.jandex.DotName;
+
+import io.quarkiverse.amazon.bedrockruntime.runtime.BedrockRuntimeBuildTimeConfig;
+import io.quarkiverse.amazon.bedrockruntime.runtime.BedrockRuntimeRecorder;
+import io.quarkiverse.amazon.common.deployment.AbstractAmazonServiceProcessor;
+import io.quarkiverse.amazon.common.deployment.AmazonClientExtensionBuildItem;
+import io.quarkiverse.amazon.common.deployment.AmazonClientExtensionBuilderInstanceBuildItem;
+import io.quarkiverse.amazon.common.deployment.RequireAmazonClientInjectionBuildItem;
+import io.quarkiverse.amazon.common.runtime.HasSdkBuildTimeConfig;
+import io.quarkiverse.amazon.common.runtime.HasTransportBuildTimeConfig;
+import io.quarkus.deployment.annotations.BuildProducer;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.annotations.ExecutionTime;
+import io.quarkus.deployment.annotations.Record;
+import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeAsyncClient;
+import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeAsyncClientBuilder;
+import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeClient;
+import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeClientBuilder;
+
+public class BedrockRuntimeProcessor extends AbstractAmazonServiceProcessor {
+
+    private static final String AMAZON_CLIENT_NAME = "amazon-sdk-bedrockruntime";
+
+    BedrockRuntimeBuildTimeConfig buildTimeConfig;
+
+    @Override
+    protected String amazonServiceClientName() {
+        return AMAZON_CLIENT_NAME;
+    }
+
+    @Override
+    protected String configName() {
+        return "bedrockruntime";
+    }
+
+    @Override
+    protected DotName syncClientName() {
+        return DotName.createSimple(BedrockRuntimeClient.class.getName());
+    }
+
+    @Override
+    protected Class<?> syncClientBuilderClass() {
+        return BedrockRuntimeClientBuilder.class;
+    }
+
+    @Override
+    protected DotName asyncClientName() {
+        return DotName.createSimple(BedrockRuntimeAsyncClient.class.getName());
+    }
+
+    @Override
+    protected Class<?> asyncClientBuilderClass() {
+        return BedrockRuntimeAsyncClientBuilder.class;
+    }
+
+    @Override
+    protected String builtinInterceptorsPath() {
+        return "software/amazon/awssdk/services/bedrockruntime/execution.interceptors";
+    }
+
+    @Override
+    protected HasTransportBuildTimeConfig transportBuildTimeConfig() {
+        return buildTimeConfig;
+    }
+
+    @Override
+    protected HasSdkBuildTimeConfig sdkBuildTimeConfig() {
+        return buildTimeConfig;
+    }
+
+    @BuildStep
+    @io.quarkus.deployment.annotations.Record(ExecutionTime.RUNTIME_INIT)
+    void createBuilders(BedrockRuntimeRecorder recorder, List<RequireAmazonClientInjectionBuildItem> amazonClientInjections,
+            BuildProducer<AmazonClientExtensionBuilderInstanceBuildItem> builderIstances) {
+        createExtensionBuilders(recorder, amazonClientInjections, builderIstances);
+    }
+
+    @BuildStep
+    @Record(ExecutionTime.RUNTIME_INIT)
+    void setup(
+            BedrockRuntimeRecorder recorder,
+            BuildProducer<AmazonClientExtensionBuildItem> amazonExtensions) {
+
+        setupExtension(recorder, amazonExtensions);
+    }
+
+}
