+++ bedrockruntime/deployment/src/test/java/io/quarkiverse/amazon/bedrockruntime/deployment/BedrockRuntimeClientFullConfigTest.java
