+++ bedrockruntime/deployment/src/test/resources/full-config.properties
@@ -1,0 +1,10 @@
+quarkus.bedrockruntime.endpoint-override=http://localhost:9090
+
+quarkus.bedrockruntime.aws.region=us-east-2
+quarkus.bedrockruntime.aws.credentials.type=static
+quarkus.bedrockruntime.aws.credentials.static-provider.access-key-id=test-key
+quarkus.bedrockruntime.aws.credentials.static-provider.secret-access-key=test-secret
+
+quarkus.bedrockruntime.sync-client.type = url
+quarkus.bedrockruntime.sync-client.connection-timeout = 0.100S
+quarkus.bedrockruntime.sync-client.socket-timeout = 0.100S
