+++ bedrockruntime/runtime/src/main/java/io/quarkiverse/amazon/bedrockruntime/runtime/BedrockRuntimeRecorder.java
