+++ devservices/bedrockruntime/src/main/java/io/quarkiverse/amazon/devservices/ecr/BedrockRuntimeDevServicesProcessor.java
@@ -2,15 +2,20 @@ package io.quarkiverse.amazon.devservices.ecr;
 
 import org.testcontainers.containers.localstack.LocalStackContainer.EnabledService;
 
-import io.quarkiverse.amazon.bedrock.BedrockBuildTimeConfig;
+import io.quarkiverse.amazon.bedrockruntime.runtime.BedrockRuntimeBuildTimeConfig;
 import io.quarkiverse.amazon.common.deployment.spi.AbstractDevServicesLocalStackProcessor;
 import io.quarkiverse.amazon.common.deployment.spi.DevServicesLocalStackProviderBuildItem;
 import io.quarkus.deployment.annotations.BuildStep;
 
-public class BedrockDevServicesProcessor extends AbstractDevServicesLocalStackProcessor {
+public class BedrockRuntimeDevServicesProcessor extends AbstractDevServicesLocalStackProcessor {
 
     @BuildStep
-    DevServicesLocalStackProviderBuildItem setupBedrock(BedrockBuildTimeConfig clientBuildTimeConfig) {
+    DevServicesLocalStackProviderBuildItem setupBedrockRuntime(BedrockRuntimeBuildTimeConfig clientBuildTimeConfig) {
         return this.setup(EnabledService.named("bedrock"), clientBuildTimeConfig.devservices());
+    }
+
+    @Override
+    protected String getPropertyConfigurationName(EnabledService enabledService) {
+        return "bedrockruntime";
     }
 }
