+++ devservices/pom.xml
@@ -19,7 +19,7 @@
     <modules>
         <module>acm</module>
         <module>bedrock</module>
-        <module>bedrock-runtime</module>
+        <module>bedrockruntime</module>
         <module>cognito-user-pools</module>
         <module>cloudwatch</module>
         <module>cloudwatchlogs</module>
