+++ docs/modules/ROOT/pages/amazon-bedrockruntime.adoc
@@ -1,0 +1,19 @@
+= Amazon Bedrock Runtime Client
+
+include::./includes/attributes.adoc[]
+
+Amazon Bedrock is a fully managed service that offers a choice of high-performing foundation models (FMs) from leading AI companies like AI21 Labs, Anthropic, Cohere, DeepSeek, Luma, Meta, Mistral AI, Stability AI, and Amazon through a single API, along with a broad set of capabilities you need to build generative AI applications with security, privacy, and responsible AI.
+
+You can find more information about Bedrock at https://aws.amazon.com/bedrock/[the Amazon Bedrock website].
+
+The Bedrock Runtime Client allows you to run inference using Amazon Bedrock models.
+
+NOTE: The Bedrock Runtime extension is based on https://docs.aws.amazon.com/sdk-for-java/v2/developer-guide/welcome.html[AWS Java SDK 2.x].
+It's a major rewrite of the 1.x code base that offers two programming models (Blocking & Async).
+
+The Quarkus extension supports two programming models:
+
+* Blocking access using URL Connection HTTP client (by default) or the Apache HTTP Client
+* https://docs.aws.amazon.com/sdk-for-java/v2/developer-guide/basics-async.html[Asynchronous programming] based on JDK's `CompletableFuture` objects and the Netty HTTP client (by default) or the AWS CRT-based HTTP client
+
+include::./includes/quarkus-amazon-bedrockruntime.adoc[]
