+++ integration-tests/pom.xml
@@ -38,7 +38,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.amazonservices</groupId>
-            <artifactId>quarkus-amazon-bedrock-runtime</artifactId>
+            <artifactId>quarkus-amazon-bedrockruntime</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.amazonservices</groupId>
