+++ integration-tests/src/main/resources/application.properties
@@ -12,7 +12,7 @@ quarkus.dynamodbenhanced.client-extensions=software.amazon.awssdk.enhanced.dynam
 
 quarkus.acm.sync-client.type=${sync-client.type}
 quarkus.bedrock.sync-client.type=${sync-client.type}
-quarkus.bedrock-runtime.sync-client.type=${sync-client.type}
+quarkus.bedrockruntime.sync-client.type=${sync-client.type}
 quarkus.cloudwatch.sync-client.type=${sync-client.type}
 quarkus.cloudwatchlogs.sync-client.type=${sync-client.type}
 quarkus.cognito-user-pools.sync-client.type=${sync-client.type}
@@ -32,7 +32,7 @@ quarkus.sts.sync-client.type=${sync-client.type}
 quarkus.sfn.sync-client.type=${sync-client.type}
 quarkus.acm.async-client.type=${async-client.type}
 quarkus.bedrock.async-client.type=${async-client.type}
-quarkus.bedrock-runtime.async-client.type=${async-client.type}
+quarkus.bedrockruntime.async-client.type=${async-client.type}
 quarkus.cloudwatch.async-client.type=${async-client.type}
 quarkus.cloudwatchlogs.async-client.type=${async-client.type}
 quarkus.cognito-user-pools.async-client.type=${async-client.type}
